clase4/mi-microservicio/service1/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
psycopg2-binary==2.9.11
pydantic==2.12.0